Clean up + refactoring

+++ nought and crosses/src/board.cpp
@@ -1,19 +1,19 @@
 #include "board.hpp"
 
-glm::vec2 cell_index_to_top_left_position(p6::Context& ctx, float nb_rows, float nb_columns, cellIndex index)
+glm::vec2 cell_index_to_top_left_position(p6::Context& ctx, int nb_rows, int nb_columns, cellIndex index)
 {
-    float xRadius = 2.f * ctx.aspect_ratio() / nb_columns;
-    float yRadius = 2.f / nb_rows;
+    float xRadius = 2.f * ctx.aspect_ratio() / static_cast<float>(nb_columns);
+    float yRadius = 2.f / static_cast<float>(nb_rows);
     float posX    = -ctx.aspect_ratio() + index.x * xRadius;
     float posY    = 1 - index.y * yRadius;
     return glm::vec2(posX, posY);
 }
 
-std::optional<cellIndex> mouse_position_to_index(p6::Context& ctx, float nb_rows, float nb_columns)
+std::optional<cellIndex> mouse_position_to_index(p6::Context& ctx, int nb_rows, int nb_columns)
 {
     if (ctx.mouse_is_in_window()) {
-        float xRadius = 2.f * ctx.aspect_ratio() / nb_columns;
-        float yRadius = 2.f / nb_rows;
+        float xRadius = 2.f * ctx.aspect_ratio() / static_cast<float>(nb_columns);
+        float yRadius = 2.f / static_cast<float>(nb_rows);
 
         int x = static_cast<int>((ctx.aspect_ratio() + ctx.mouse().x) / xRadius);
         int y = static_cast<int>((1 - ctx.mouse().y) / yRadius);
@@ -23,18 +23,18 @@ std::optional<cellIndex> mouse_position_to_index(p6::Context& ctx, float nb_rows
     return {};
 }
 
-void draw_board(p6::Context& ctx, float nb_rows, float nb_columns)
+void draw_board(p6::Context& ctx, int nb_rows, int nb_columns)
 {
     ctx.use_fill  = true;
     ctx.fill      = {0.5f, 0.5f, 0.5f, 1.f};
     ctx.stroke    = {1.f, 1.f, 1.f, 1.f};
     float xTile   = -ctx.aspect_ratio();
     float yTile   = 1;
-    float xRadius = 2.f * ctx.aspect_ratio() / nb_columns;
-    float yRadius = 2.f / nb_rows;
+    float xRadius = 2.f * ctx.aspect_ratio() / static_cast<float>(nb_columns);
+    float yRadius = 2.f / static_cast<float>(nb_rows);
 
-    for (int y = 0; y < static_cast<int>(nb_rows); y++) {
-        for (int x = 0; x < static_cast<int>(nb_columns); x++) {
+    for (int y = 0; y < nb_rows; y++) {
+        for (int x = 0; x < nb_columns; x++) {
             ctx.rectangle(p6::TopLeftCorner{xTile, yTile}, p6::Radii{xRadius, yRadius});
             xTile += xRadius;
         }
